Create profile by username page
