+++ src/app/(auth)/profile/[username]/page.tsx
@@ -1,0 +1,35 @@
+"use client"
+
+import ProfilePage from "../components/ProfilePage";
+
+import { useUserByUsename } from "@/hooks/useUserProfile";
+import { useParams } from 'next/navigation';
+import { useSession } from "next-auth/react";
+import LoadingState from "@/components/LoadingState";
+
+export default function ProfileByUserName() {
+  const params = useParams();
+  const username = params?.username.toString();
+  const formattedUsername = username.replace(/_/g, " ")
+
+  const { userByUsername } = useUserByUsename(formattedUsername.toString());
+  const { data: session } = useSession();
+
+  const formattedUsernameToLower = formattedUsername?.toLowerCase().trim();
+  const loggedUserNameToLower = session?.user?.name?.toLowerCase().trim() || "";
+  const isSameUser = formattedUsernameToLower === loggedUserNameToLower;
+
+  if (!userByUsername) {
+    return(
+      <LoadingState/>
+    )
+  }
+
+  return(
+    <ProfilePage 
+      isSameUser={isSameUser} 
+      profile={userByUsername}
+    >
+    </ProfilePage>
+  )
+}
