+++ src/hooks/useUserProfile.ts
@@ -31,3 +31,32 @@ export function useUserProfile() {
 
   return { userProfile, isLoading, error };
 }
+
+export function useUserByUsename(search?: string) {
+  const { data: session } = useSession();
+  const [userByUsername, setUserByUsername] = useState<UserProfile | null>(null);
+  const [isLoading, setIsLoading] = useState(true);
+  const [error, setError] = useState<string | null>(null);
+  useEffect(() => {
+    const fetchAllUsers = async () => {
+      if (session?.user?.id && session?.user?.token) {
+        try {
+          const data = await userService.fetchAllUsers(
+            session.user.token,
+            search
+          );
+          setUserByUsername(data);
+        } catch (error) {
+          console.error("Error fetching user by user name:", error);
+          setError(error.message);
+        } finally {
+          setIsLoading(false);
+        }
+      }
+    };
+
+    fetchAllUsers();
+  }, [session]);
+
+  return { userByUsername, isLoading, error };
+}
