+++ src/services/userService.ts
@@ -2,9 +2,17 @@ import axios from "axios";
 
 export const userService = {
   async fetchUserProfile(userId: string, token: string) {
-    const response = await axios.get(`/api/users`, {
+    const response = await axios.get(`/api/users/${userId}`, {
+      headers: {
+        Authorization: `Token ${token}`,
+      },
+    });
+    return response.data;
+  },
+  async fetchAllUsers(token: string, search?: string) {
+    const response = await axios.get(`/api/users/`, {
       params: {
-        userId,
+        search,
       },
       headers: {
         Authorization: `Token ${token}`,
